requirements.txt
opencv-python>=4.9.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-image>=0.22.0
pytest>=8.0.0
reportlab>=4.0.0
fastapi>=0.100.0
uvicorn>=0.24.0
python-multipart>=0.0.6
